+++ inst/doc/model_builder.R
@@ -1,0 +1,4 @@
+## ----echo=FALSE, message=FALSE, warning=FALSE, eval=TRUE,  style="max-height: 100px;", comment=""----
+#yaml= file.path(system.file(package="ruminate"), "templates", "MB.yaml")
+cat(readLines(file.path(system.file(package="ruminate"), "templates", "MB.yaml")), sep="\n")
+